sgav_backend/win_requirements.txt
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==22.1.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.0.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.2
Deprecated==1.2.14
diskcache==5.6.3
distro==1.8.0
dnspython==1.16.0
docopt==0.6.2
docx2txt==0.8
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
future==1.0.0
gast==0.4.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.19.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
kubernetes==28.1.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.31
libclang==18.1.1
llama_cpp_python==0.2.18
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==5.2.0
mypy-extensions==1.0.0
networkx==2.6.3
nltk==3.8.1
numpy==1.23.5
# nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
overrides==7.4.0
packaging==23.2
Pillow==10.1.0
posthog==3.0.2
prompt-toolkit==3.0.28
protobuf==3.19.6
psutil==5.9.6
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydot==1.4.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.1.2
pypdf==3.17.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.9.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rasa-sdk==3.5.1
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.52
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.6.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
# triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.10.0
tzlocal==5.2
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
# uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.6.4
websockets==10.4
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
